full integration & implmentation of game engine logic

+++ core/src/com/mygdx/game/managers/CollisionManager.java
@@ -8,6 +8,7 @@ import com.mygdx.game.models.Entity;
 
 public class CollisionManager {
 	private List<Entity> collidableList;
+	private List<Entity> updatedEntityList;
 	
 	public CollisionManager(List<Entity> entityList) {
 		
@@ -18,14 +19,21 @@ public class CollisionManager {
 		}
 	}
 	
+    public void update(List<Entity> entityList) {
+    	updatedEntityList = new ArrayList<Entity>();
+        for (Entity entity : entityList) {
+            if (entity.isCollidable()) {
+                updatedEntityList.add(entity);
+            }
+        }
+        this.collidableList = updatedEntityList;
+    }
+	
 	 public void checkCollisions(EntityManager entityManager) {
 		 for (int i = 0; i < collidableList.size() - 1; i++) {
 	            for (int j = i + 1; j < collidableList.size(); j++) {
 
 	                if (collidableList.get(i).collidesWith(collidableList.get(j))) {
-//	                	collidableList.get(i).onCollision(collidableList.get(j));
-//	                	collidableList.get(j).onCollision(collidableList.get(i));
-	                	System.out.println("Collision Detected");
 	                	handleCollisions(entityManager,collidableList.get(i),collidableList.get(j));
 	                	
 	                }
@@ -33,16 +41,23 @@ public class CollisionManager {
 	        }
 	    }
 	 public void handleCollisions(EntityManager entityManager, Entity x, Entity y) {
-		 if(x.isAiControl()) {
-			 collidableList.remove(x);	//update collidableList to remove entity
-			 entityManager.removeEntity(x);	//update entityManager to remove entity
-			 System.out.println("AI Entity Removed");
+		 if(x.isAiControl() && y.isAiControl()) {
+//			 System.out.println("Collision Between 2 AIs - Allowed");
+		 }else if(!x.isAiControl() && !y.isAiControl()){
+//			 System.out.println("Collision Between 2 Players - Allowed");
+		 }else {
+			 if(x.isAiControl()) {
+				 collidableList.remove(x);	//update collidableList to remove entity
+				 entityManager.removeEntity(x);	//update entityManager to remove entity
+				 System.out.println("Collision Between Player & AI - AI Entity Removed");
+			 }
+			 else {
+				 collidableList.remove(y);	//update collidableList to remove entity
+				 entityManager.removeEntity(y);	//update entityManager to remove entity
+				 System.out.println("Collision Between Player & AI - AI Entity Removed");
+			 }
 		 }
-		 else {
-			 collidableList.remove(y);	//update collidableList to remove entity
-			 entityManager.removeEntity(y);	//update entityManager to remove entity
-			 System.out.println("AI Entity Removed");
-		 }
+		 
 		 
      	//else entityManager.removeEntity(y);
